backend/requirements.txt
fastapi
gunicorn
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
fastapi-mail
python-dotenv
psycopg2-binary